+++ Lab_1/Makefile
@@ -1,0 +1,4 @@
+all:
+	gcc -c archiver.c unarchiver.c
+	gcc -c program.c
+	gcc -o arch program.o archiver.o unarchiver.o
