+++ Lab_1/lib.h
@@ -1,0 +1,3 @@
+void RecursiveArchive(char *, int);
+void Archive(char *, char *);
+void Unarchive(char *, char *);
